fix(typo/spell): :pencil2: change ext hook file to `.ts`
